pyproject.toml
[project]
name = "property_scanner"
version = "0.1.0"
description = "End-to-end multi-agent system for discovering and evaluating real-estate opportunities."
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.0",
  "pydantic-settings>=2.12.0",
  "hydra-core>=1.3",
  "requests>=2.31",
  "beautifulsoup4>=4.12",
  "structlog>=23.1",
  "pyyaml>=6.0",
  "pandas>=2.0",
  "duckdb>=0.9",
  "sqlalchemy>=2.0",
  "tenacity>=8.2",
  "geopy>=2.4",
  "playwright>=1.40.0",
  "playwright-stealth>=1.0.6",
  "streamlit>=1.30.0",
  "curl_cffi>=0.7.3",
  "plotly>=5.18.0",
  "sentence-transformers>=2.2.2",
  "numpy>=1.26",
  "lightgbm>=4.1",
  "scikit-learn>=1.3",
  "xgboost>=2.0",
  "langgraph>=0.2",
  "langchain-core>=0.3",
  "langchain-openai>=0.2",
  "langchain-community>=0.3",
  "litellm>=1.44",
  "instructor>=1.6",
  "aiofiles>=25.1",
  "aiohttp>=3.9",
  "websockets>=14.0",
  "typing-extensions>=4.14",
  "prefect>=3.6",
  "fastapi>=0.116",
  "uvicorn>=0.35",
  "polars>=0.20",
  "lancedb>=0.10",
]

[dependency-groups]
dev = [
  "pytest>=8.4",
  "black>=23.0",
  "isort>=5.0",
  "mypy>=1.0",
  "jupyter>=1.0",
]

[tool.poetry]
name = "property_scanner"
version = "0.1.0"
description = "End-to-end multi-agent system for discovering and evaluating real-estate opportunities."
authors = ["Antigravity <bot@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.0"
pydantic-settings = ">=2.12.0"
hydra-core = "^1.3"
requests = "^2.31"
beautifulsoup4 = "^4.12"
structlog = "^23.1"
pyyaml = "^6.0"
pandas = "^2.0"
duckdb = "^0.9"
sqlalchemy = "^2.0"
tenacity = "^8.2"  # For retries
geopy = "^2.4"
playwright = "^1.40.0"
playwright-stealth = "^1.0.6"
streamlit = "^1.30.0" # For dynamic sites like Idealista
curl_cffi = "^0.7.3"
plotly = "^5.18.0"
sentence-transformers = "^2.2.2"
numpy = "^1.26"
lightgbm = "^4.1"
scikit-learn = "^1.3"
xgboost = "^2.0"
langgraph = "^0.2"
langchain-core = "^0.3"
langchain-openai = "^0.2"
langchain-community = "^0.3"
litellm = "^1.44"
instructor = "^1.6"
aiofiles = "^25.1"
aiohttp = "^3.9"
websockets = "^14.0"
typing-extensions = "^4.14"
prefect = "^3.6"
fastapi = "^0.116"
uvicorn = "^0.35"
polars = "^0.20"
lancedb = "^0.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
isort = "^5.0"
mypy = "^1.0"
jupyter = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
